docs/For Reference/AD_Auth/backend/requirements.txt
# CloudEstimator Backend Requirements
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
ldap3==2.9.1

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client (for pricing APIs)
httpx==0.26.0
aiohttp==3.9.1

# Export Generation
python-docx==1.1.0
weasyprint==60.2
openpyxl==3.1.2
jinja2==3.1.3
Pillow==10.2.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Logging
structlog==24.1.0
